+++ src/pages/certificates/create-certificate/tests/create-certificate.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import CreateCertificate from '../create-certificate';
+import CreateCertificate from '..';
 
 describe('test CreateCertificate component', () => {
   it('should change quantity in appropriate h5', () => {
